refactor: reshape the structure of schema-related DAGs tree
