+++ schema/consolidation/README.md
@@ -1,0 +1,12 @@
+# Documentation
+
+## schema_consolidation
+
+| Information | Valeur |
+| -------- | -------- |
+| Fichiers sources     | `DAG_schema_consolidation.py`     |
+| Description | Ce traitement permet de consolider les ressources de data.gouv.fr liées à un schéma en un fichier global et de labéliser les données en lien avec les schémas de données. |
+| Fréquence de mise à jour | Quotidienne |
+| Données sources | Données de data.gouv.fr et de schema.data.gouv.fr |
+| Données de sorties | Jeux de données sur data.gouv.fr et repository de schema.data.gouv.fr |
+| Channel Mattermost d'information | ~datagouv-schema-activites |
