+++ schema/consolidation/schemas_consolidation.md
