+++ schema/recommendations/README.md
@@ -1,0 +1,12 @@
+# Documentation
+
+## schema_recommendations
+
+| Information | Valeur |
+| -------- | -------- |
+| Fichiers sources     | `DAG_schema_recommendations.py`     |
+| Description | Ce traitement permet de générer des recommandations entre jeux de données sur data.gouv.fr labelisés avec des schémas. Les recommandations consistent en un lien visible sur la page d'un jeu de donnée data.gouv.fr vers le fichier national lié au schéma des ressource. |
+| Fréquence de mise à jour | Quotidienne |
+| Données sources | Données de data.gouv.fr et de schema.data.gouv.fr |
+| Données de sorties | Fichiers de métadonnées |
+| Channel Mattermost d'information | Aucun |
